Video pausing with proper audio synchronization; render a timebar when hovering video
The video can now be paused using the Space key. When played, it will also properly resume the audio by making a request for a part of the audio file (one which starts from a custom position), which the server now supports with the help of FFmpeg.

A timebar is now also rendered at the bottom of the video, when the mouse is hovering over it, or when it's paused. You cannot currently interact with it, however. That is to-be-done.