Use "HTTP " prefix to specify an HTTP request in Scratch
The Scratch userscript hack now ensures that a translate/speak request starts with "HTTP " before redirecting the request to the custom server. This enables other Scratch project that utilize translation/synth to function properly with the userscript enabled.

The server always checks and removes this prefix, if available. This allows for compatibility with any other Scratch HTTP GET request hacks that might be utilized in the future (because since the userscript removes the "HTTP " prefix, other hacks are not guaranteed to be able to do that).