Use FFmpeg (with FFprobe) to get video frames, replacing OpenCV
The dependency on OpenCV is now removed. FFmpeg (with FFprobe) now takes care of getting video frames to be converted to hex RGB pixel data.

This method is way faster, making the request response time around 200-300ms, as opposed to the previous method of doing `next()` to skip frames with OpenCV, where it was seconds long and getting longer the further frames are requested from the video.